Fix function signature for integer variant of round
Fixes a bug introduced when the the integer type was added, which was
exacerbated by our limited type coercion system. As only a handful of
variants of our math functions exist, calls to the function
round(tinyint, tinyint) would fail due to unresolvable ambiguities
between round(integer, integer) or round(bigint, bigint).

+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/MathFunctions.java
@@ -439,7 +439,7 @@ public final class MathFunctions
     @Description("round to nearest integer")
     @ScalarFunction("round")
     @SqlType(StandardTypes.INTEGER)
-    public static long roundInteger(@SqlType(StandardTypes.INTEGER) long num, @SqlType(StandardTypes.INTEGER) long decimals)
+    public static long roundInteger(@SqlType(StandardTypes.INTEGER) long num, @SqlType(StandardTypes.BIGINT) long decimals)
     {
         return num;
     }
